requirements.txt
streamlit
openai
python-docxv